+++ lambda_aws/lambdas.py
@@ -1,0 +1,37 @@
+import boto3,json
+import os
+
+class LambdaDishPlus:
+    def deleteSuscriberSes(id_customer,email):
+        try:
+            payload = {
+                "id_customer": id_customer,
+                "email": email
+            }
+            lambda_client = boto3.client('lambda',region_name="us-east-1")
+            lambda_url = os.environ.get("LAMBDA_DELETE_SES")
+            response_json = lambda_client.invoke(FunctionName = lambda_url, InvocationType = "RequestResponse",Payload = json.dumps(payload))
+            response = json.loads(response_json['Payload'].read())
+            return response
+            
+        except Exception as exc:
+            return exc
+
+    def deleteSuscriber(email):
+            
+        try:
+            payload = {
+                "method": "Post",
+                "sourceIP": os.environ.get("SOURCE"),
+                "body": {
+                        "email": email
+                }
+            }
+            lambda_client = boto3.client('lambda',region_name="us-east-1")
+            lambda_url = os.environ.get("LAMBDA_DELETE_SUSCRIBER")
+            response_json = lambda_client.invoke(FunctionName = lambda_url, InvocationType = "RequestResponse",Payload = json.dumps(payload))
+            response = json.loads(response_json['Payload'].read())
+            return response
+            
+        except Exception as exc:
+            return exc
