+++ utils/cognitos.py
@@ -25,38 +25,3 @@ class CognitoDishPlus:
         
         return response
     
-    def deleteSuscriberSes(id_customer,email):
-        
-        try:
-            payload = {
-                "id_customer": id_customer,
-                "email": email
-            }
-            lambda_client = boto3.client('lambda',region_name=os.environ.getenv("REGION_NAME"))
-            lambda_url = os.environ.getenv("LAMBDA_DELETE_SES")
-            response_json = lambda_client.invoke(FunctionName = lambda_url, InvocationType = "RequestResponse",Payload = json.dumps(payload))
-            response = json.loads(response_json['Payload'].read())
-            return response
-        
-        except Exception as exc:
-            return exc
-
-    def deleteSuscriber(email):
-        
-        try:
-            payload = {
-                "method": "Post",
-                "sourceIP": os.environ.getenv("SOURCE"),
-                "body": {
-                        "email": email
-                }
-            }
-            lambda_client = boto3.client('lambda',region_name=os.environ.getenv("REGION_NAME"))
-            lambda_url = os.environ.getenv("LAMBDA_DELETE_SUSCRIBER")
-            response_json = lambda_client.invoke(FunctionName = lambda_url, InvocationType = "RequestResponse",Payload = json.dumps(payload))
-            response = json.loads(response_json['Payload'].read())
-            return response
-        
-        except Exception as exc:
-            return exc
-
